Replaced Jquery UI with Foundation. Because 2018
