+++ Gemfile
@@ -22,7 +22,7 @@ gem 'uglifier', '>= 1.3.0'
 # Use CoffeeScript for .coffee assets and views
 gem 'coffee-rails', '~> 4.2'
 # Turbolinks makes navigating your web application faster. Read more: https://github.com/turbolinks/turbolinks
-gem 'turbolinks', '~> 5'
+# gem 'turbolinks', '~> 5'
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
 gem 'jbuilder', '~> 2.5'
 # Use Redis adapter to run Action Cable in production
@@ -58,7 +58,7 @@ gem 'sys-filesystem'
 gem 'codemirror-rails'
 
 gem 'jquery-rails'
-gem 'jquery-ui-rails'
+gem 'foundation-rails'
 
 gem 'git'
