--- app/assets/stylesheets/dashboard.scss
@@ -1,75 +1,0 @@
-html, body{
-  background: #272822;
-  color: #CCCCCC;
-  font-family: "Trebuchet MS", sans-serif;
-  height: 100%;
-}
-
-#tabs{
-  height: 80%;
-}
-
-div[id^=tab]{
-  height: 100%;
-  background: #272822;
-  color: #CCCCCC;
-}
-
-input[type=text]{
-  width: 50%;
-}
-
-.CodeMirror{
-  height: 75% !important;
-  width: 80% !important;
-  display: inline-block !important;
-}
-
-.column-3{
-  column-count: 3;
-}
-
-.dataTables_wrapper{
-  background: #8f8f8f;
-  padding: 10px;
-}
-
-.odd{
-  color: #333333;
-  background: #bbbbbb !important;
-}
-
-.even{
-  color: #333333;
-}
-
-li a{
-  color:#CCCCCC !important;
-}
-
-li a:visited{
-  color:#CCCCCC;
-}
-
-li a:hover{
-  color:#CCCCCC;
-}
-
-.switch-file-browser{
-  float: right;
-  width: 15%;
-  border: solid 1px #CCCCCC;
-  padding: 0 10px 10px 10px;
-  -webkit-border-radius: 5px;
-  -moz-border-radius: 5px;
-  border-radius: 5px;
-}
-
-.switch-file-browser ul{
-  padding-left: 15px;
-}
-
-.active-file, .local-path{
-  color: #007FFF !important;
-}
-
