+++ app/assets/stylesheets/studio.scss
@@ -1,0 +1,2 @@
+.CodeMirror{
+}
