+++ app/helpers/application_helper.rb
@@ -1,2 +1,5 @@
 module ApplicationHelper
+  def active_page(path)
+    'active' if request.url.include?(path)
+  end
 end
