--- app/helpers/dashboard_helper.rb
@@ -1,2 +1,0 @@
-module DashboardHelper
-end
