+++ app/helpers/studio_helper.rb
@@ -1,0 +1,2 @@
+module StudioHelper
+end
