--- app/views/dashboard/index.html.erb
@@ -1,14 +1,0 @@
-<h2>BashBunny Studio</h2>
-<%= render 'dashboard/partials/bunny_mount_point' %>
-<p>
-  <hr/>
-</p>
-<div id="tabs">
-  <%= render 'dashboard/partials/tab_menu' %>
-  <%= render 'dashboard/partials/switch', switch_position: 1 %>
-  <%= render 'dashboard/partials/switch', switch_position: 2 %>
-  <%= render 'dashboard/partials/extensions' %>
-  <%= render 'dashboard/partials/payloads' %>
-  <%= render 'dashboard/partials/git' %>
-  <%= render 'dashboard/partials/debug' %>
-</div>
