--- app/views/dashboard/partials/_debug.html.erb
@@ -1,11 +1,0 @@
-<div id="tabs-debug">
-  <% if @bash_bunny_path.include?('/') %>
-      <ul>
-        <% path = '/debug' %>
-        <% Dir.foreach("#{@bash_bunny_path}#{path}") do |debug_file| %>
-            <% next if debug_file == '.' || debug_file == '..' %>
-            <li><a href="/dashboard/raw_file?path=<% path %>&file=<%= debug_file %>" target="_blank"><%= debug_file %></a></li>
-        <% end %>
-      </ul>
-  <% end %>
-</div>
