--- app/views/dashboard/partials/_git.html.erb
@@ -1,16 +1,0 @@
-<div id="tabs-git">
-  <h4 class="clone-setting-repo clone-setting">Repo:</h4>
-  <input type="text" id="git-repo" class="clone-setting">
-  <h4 class="pull-setting">Local Location:</h4>
-  <span class="pull-setting"><%= "#{Rails.root}/#{@local_repo_path}" %></span>
-  <p id="git-action-buttons">
-    <button class="git sync clone-setting" onclick="cloneRepo()">Clone (90 Sec)</button>
-    <input type="text" id="git-command" class="pull-setting" placeholder="e.g. checkout develop" />
-    <button class="git pull pull-setting" onclick="gitCommand()">Run Command</button>
-    <p id="git-command-response" class="pull-setting">
-    </p>
-    <p>
-      <button class="git delete pull-setting" onclick="deleteRepo()">Delete</button>
-    </p>
-  </p>
-</div>
