--- app/views/dashboard/partials/_switch.html.erb
@@ -1,11 +1,0 @@
-<% switch_position = local_assigns[:switch_position].to_s %>
-<div id="tabs-switch-<%= switch_position %>">
-  <p>
-    <button class="switch-<%= switch_position %> save" onclick="writePayload(<%= switch_position %>)">Save</button>
-  </p>
-  <p class="switch-<%= switch_position %> local-path"></p>
-  <hr/>
-  <textarea class="switch-<%= switch_position %> payload-editor">
-  </textarea>
-  <%= render 'dashboard/partials/switch_file_browser', switch_position: switch_position %>
-</div>
