--- app/views/dashboard/partials/_tab_menu.html.erb
@@ -1,8 +1,0 @@
-<ul>
-  <li><a href="#tabs-switch-1">Switch 1</a></li>
-  <li><a href="#tabs-switch-2">Switch 2</a></li>
-  <li><a href="#tabs-extensions">Extensions</a></li>
-  <li><a href="#tabs-payloads">Payloads</a></li>
-  <li><a href="#tabs-git">Git</a></li>
-  <li><a href="#tabs-debug">Debug</a></li>
-</ul>
