+++ app/views/shared/_bunny_mount_point.html.erb
@@ -1,0 +1,1 @@
+
