+++ app/views/studio/debug.html.erb
@@ -1,0 +1,10 @@
+<%= javascript_include_tag 'debug' %>
+<% if bunny_mounted? %>
+    <ul>
+      <% path = '/debug' %>
+      <% Dir.foreach("#{bunny_path}#{path}") do |debug_file| %>
+          <% next if debug_file == '.' || debug_file == '..' %>
+          <li><a href="/dashboard/raw_file?path=<% path %>&file=<%= debug_file %>" target="_blank"><%= debug_file %></a></li>
+      <% end %>
+    </ul>
+<% end %>
