+++ app/views/studio/learn.html.erb
@@ -1,0 +1,1 @@
+<%= javascript_include_tag 'learn' %>
