+++ app/views/studio/payloads.html.erb
@@ -1,0 +1,30 @@
+<%= javascript_include_tag 'payloads' %>
+<table class="data-table" width="100%">
+  <thead>
+  <td>Name</td>
+  <td>Category</td>
+  <td class="no-sort">Copy To</td>
+  </thead>
+  <tbody>
+  <% if File.exist?("#{local_repo_path}/payloads/extensions") %>
+      <% Dir.foreach("#{local_repo_path}/payloads/library") do |category| %>
+          <% next if category.include?('.') %>
+          <% Dir.foreach("#{local_repo_path}/payloads/library/#{category}") do |payload| %>
+              <% next if payload.include?('.') %>
+              <tr>
+                <td><%= payload %></td>
+                <td><%= category.humanize %></td>
+                <td>
+                  <button onclick="copyPayload(1, '<%= "#{local_repo_path}/payloads/library/#{category}/#{payload}" %>')">Switch
+                    1
+                  </button>
+                  <button onclick="copyPayload(2, '<%= "#{local_repo_path}/payloads/library/#{category}/#{payload}" %>')">Switch
+                    2
+                  </button>
+                </td>
+              </tr>
+          <% end %>
+      <% end %>
+  <% end %>
+  </tbody>
+</table>
