+++ app/views/studio/switch1.html.erb
@@ -1,0 +1,2 @@
+<%= javascript_include_tag 'switch' %>
+<%= render 'shared/switch', switch_position: 1 %>
