+++ config/routes.rb
@@ -1,16 +1,24 @@
 Rails.application.routes.draw do
-  get 'dashboard/index'
-  get 'dashboard/check_repo'
-  get 'dashboard/clone_repo'
-  get 'dashboard/git_command'
-  get 'dashboard/delete_repo'
-  get 'dashboard/eject_bunny'
-  get 'dashboard/copy_payload'
-  get 'dashboard/sync_extensions'
-  get 'dashboard/raw_file'
-  post 'dashboard/write_payload'
-  get 'dashboard/payload_script'
-  get 'dashboard/index'
+  get 'studio/switch1'
+  get 'studio/switch2'
+  get 'studio/extensions'
+  get 'studio/payloads'
+  get 'studio/loot'
+  get 'studio/debug'
+  get 'studio/git'
+  get 'studio/learn'
 
-  root 'dashboard#index'
+  get 'studio/clone_repo'
+  get 'studio/git_command'
+  get 'studio/delete_repo'
+  get 'studio/eject_bunny'
+  get 'studio/copy_payload'
+  get 'studio/sync_extensions'
+  get 'studio/raw_file'
+  get 'studio/payload_script'
+
+  post 'studio/write_payload'
+
+
+  root 'studio#switch1'
 end
